Move graph types out of graph file
